Supply Consoles no longer require cargo access to use 2: SOLDERING EDITION
See #22008
As per comments made in the last PR which was made in badwrongevilsatan web editor, soldering the cargo supply console (the one inside the cargo office, not the one in the hallway) will now remove the cargo access required to approve/deny/call the supply shuttle/return it/etc.
tl;dr current access code a shit and soldering the board has been deemed an okay compromise for a way to (temporarily) delete the access bottleneck of cargo, featuring 1 dead miner and the spare being hidden somewhere.

[controversial][balance]

:cl:
 * tweak: Soldering the Cargo Supply Console board will remove the access requirement to approve/deny orders and to call/return the supply shuttle.