Removes the NT12 and Automag magazines from the Ammolathe. (#22084)
It's okay when I decry ~~griff~~ FUN that pertains to other departments, but not when others do it to my departments.

On a slightly more serious note, I should have paid more attention to the Ammolathe PR, because good lord this is a case of Vectors being borked and the c*der behind it powercreeping his content intentionally. For the record, you can currently take magazines meant for guns that are intended to be adminbus/wizard/traitor-only (NT12 and Automag) and load them into a Vector. Said vector just so happens to be semi-auto and have a really nice burst-fire feature. You know what regular shotguns don't have? Semi-auto and burst-fire. I lurv me some 120 brute in half a second (quoting Sonix here). The only shotgun that doesn't quite get rendered obsolete by the Vector is the Combat Shotgun, because of its 8-round capacity.
Why the Auytomag and NT12 more specifically as opposed to the other magazines? That's simple. They both use calibers that are readily available to the crew/the people that would handle a Vector (Security has boxes of shotgun ammo and can print more at the ammolathe, .357 can be printed at an autolathe). Why not the other magazines? Because outside of 9mm, they're practically unobtainium calibers on the station.

tl;dr Vectors are borked and Ammolathe highlighted their borkedness even more.