Wizard Doppelgangers now have 40 HP.
I didn't quite die, but it turns out that unstunnable mobs that can almost chainstun people and deal mild damage at the same time, all the while being spammable on a 30/20/10 seconds cooldown is a bit iffy. All that done with a nice 100 HP. These things might be the tankiest, spammable & incosequential pest I've seen in a while.