MoMMIs can use microwaves
It is improving the station to cook food if the station has no freshly microwaved food.
MoMMI's can use chemical dispensers and ChemicalMasters, but can't figure out how to use the bartending equipment?
I understand borgs other than service not being able to use the microwave due to it not being updated to allow silicon control, but MoMMI's have an arm and can do other things with it, why not turn on a microwave? They can plant and harvest crops, why can't they turn on a microwave?
I will stock the medbay fridge every round as a chemist MoMMI every round until this gets merged and once again I can be a bartender MoMMI to the fullest extent.
Fight me balance fascists.