Mob printers no longer care if you can re-enter corpse
Mob printers no longer check to see if you can re-enter corpse. This is probably sensible sanity but since it's an adminbus piece of machinery, and it breaks under some very annoying circumstances (for instance, if you run into the bluespace in Lamprey Hell) then it's preferable to just let people respawn as something more often than not.