Rogue Fabricators
It was inevitable. The Neo-Soviet MoMMI units, left for countless years to maintain a constantly degrading station, have found a new way of breaking their z-locks and programming, so as to proliferate.

The fabricator arrives through hiding within the midst of a meteor shower, colliding with the edge of an unknowing station, burrowing in like a tick, and beginning its programming cycle: Consume, Reproduce, Continue.

The end goal of a rogue fabricator is fairly open. Produce worker drones through the hijacked fabricator it is built around, produce more fabricators from the acquired metal, and break down the station until there is nothing left of use.