Add back the changeling
In the few times I've seen the changeling bused it's been a whole lot of fun to play against.   I understand that a lot of people are not a fan of the ling,  however I think the need for more antag types (we have a variety issue) warrants it's readdition. 

I think a compromise is warranted. The compromise comes in line with the the other change I'm proposing, which  is to vastly reduce the chances of getting changeling in dynamic.   If it turns out that changelings are terrible no matter what,  it won't affect too many rounds before removal. If it turns out that server culture has changed enough for ling to be fun, then we can up the dynamic rate.  If changeling is fun as a once in a while treat then we can keep it rare.