Make only one changeling spawn
Look I don't know how to code, sorry for making so many PRS.  

this is part of the "test run" if it turns out that one ling is fun we can test with more. 

I am less confident about this change, if anything it means less people get to try it for themselves.  However I understand that people are worried about the perma removal. 

If you'd rather have more lings thumbs this one down.  
.