Adds "recall" artifact effect.
When activated, the positions of all elegible objects (basically unanchored) objects around it are recorded. When the artifact deactivates after a charge time, all the recorded objects are teleported *back* to their position at the time of activation.

Objects are only sent back if the container they're in actually supports it. So no breaking the game sorry.

Also added a weakref type since I need that for this to avoid causing hard dels.