Packed power and mechanics
Changed the mechanics office to be a bit more functional using the dead space around it and now mechanics, naturally distrustful of one another, no longer spawn directly adjacent to each other. Also added a SMES to engineering to charge that powernet such as exists on all other stations. No idea when it went missing since I remember it existing before and now results in new engineers loosing the singulo all the time.