Bring back TRUE democracy (#31551)
Democracy is not the tyranny of the majority. Never has been, never will.
Democracy is about protecting the rights of the minorities while ensuring the governement works for everyone.

We MUST fight for the rights of indigenous meta-voting PoC (People of [map] choice). The future cannot be all Box, Roid, and Box again.

Weighted voting works in such a way that if, globally, 20% of people vote for Packed, Packed will be played, on average, 20% of the time. This is true fairness.
Absolute voting means that if, globally, 40% of people vote for Box, 39% for Roid and 21% for Packed, Box will be voted 100% of the time.

This is absolutely unfair and a real danger to map diversity, ~~equality, and inclusion~~.

:cl:
- rscadd: Bring back weighted voting.