Librarians now start with a book bag
Someone requested it, and I wanted to learn how to work this github doodad and code malarkey. So far it looks pretty ugly and I'm not sure how I would get it to autodump books into a shelf when clicked on it given the extent of my knowledge is copying and pasting with slight edits. But even if I can't, it still holds more books than say, a backpack, which is helpful for stocking the shelves. I'll scratch my head some more and see if I can figure something out though.

It holds 25 books(vs the backpack's 12), including basic books, spellbooks, and cult tomes. If there's some reason I overlooked that it shouldn't hold cult tomes yell at me for it and I'll fix it. I excluded bibles from it since you can store things inside of them.