Advanced Health Scanner HUD (#31930)
* Advanced Health Scanner HUD

The CMO's locker now spawns with a pair of Advanced MedHUDs. They function as regular health scanner huds with the addition of having flash protection, meaning they'll no longer be the only head without it roundstart. Only one pair spawns in the locker and they can't be scanned by the mechanic's scanner.

I wanted to add science goggles function as well but I'm too stupid to do it currently so if this gets merged I might try to PR that at a later date. I also don't entirely love the sprite so I'll probably mess with it more.

* Sprite Updates, Syndicate Scanner

* Adds science soggle effect