Makes Tackling Disarm Intent Only (#32107)
I have a pet peeve with tackling. When I want to throw a person after grabbing them aggressively, I spamclick the throwing part of it. That's what you do when fighting people in ss13, you click often and click fast. There are so many click delays to things and you want to make sure you get your inputs in there.

But this leads to a problem where you accidentally end up tackling yourself after throwing a guy, I've grown tired of this happening, so here's a one line tweak.

:cl:
 * tweak: Tackling is now exclusive to disarm intent instead of both disarm intent and grab intent