Adds Paints to Trade Outfitter
[content]

## What this does

Adds a new toolbox to the trade outfitter containing each color of paint, available for purchase for 80 credits. This does not contain canvases, a paintbrush, or a palette, as traders can produce their own wood.

## Why it's good

Adds a new avenue an artistically inclined trader without having to make a deal with their biggest competitor, cargo.

## Changelog

* rscadd: Added painter's toolbox containing paints to the trade outfitter.

* imageadd: Added sprites for the painter's toolbox.