Price update/Small sprite change
The initial thought behind the high price was to make it undesirable to try and simply flip the paints instead of actually using them but I suppose for that it just needs to be above 10 by any amount. Changed price to 40 to ensure that a trader with a poor starting roll can still purchase a junk crate as well.

Also changed the in-hand sprites border on the new toolbox to be the same color as the world sprite instead of a bright orange.