Adds Paints to Trade Outfitter (#33427)
* Adds Paints to Trade Outfitter

[content]

## What this does

Adds a new toolbox to the trade outfitter containing each color of paint, available for purchase for 80 credits. This does not contain canvases, a paintbrush, or a palette, as traders can produce their own wood.

## Why it's good

Adds a new avenue an artistically inclined trader without having to make a deal with their biggest competitor, cargo.

## Changelog

* rscadd: Added painter's toolbox containing paints to the trade outfitter.

* imageadd: Added sprites for the painter's toolbox.

* Price update/Small sprite change

The initial thought behind the high price was to make it undesirable to try and simply flip the paints instead of actually using them but I suppose for that it just needs to be above 10 by any amount. Changed price to 40 to ensure that a trader with a poor starting roll can still purchase a junk crate as well.

Also changed the in-hand sprites border on the new toolbox to be the same color as the world sprite instead of a bright orange.

* Update toolbox_ihr.dmi

Small sprite mistake corrected