Make perforate not throw straight down when clickdrag starts from soulblade's tile.
Clickdrags starting from the same tile as the soulblade, would throw
straight down regardless of where you dragged to, because a soulblade's
clickdrag throw has two parts: It first throws the soulblade to where
the clickdrag starts, then redirects the soulblade to where the
clickdrag ends. Starting the clickdrag from the soulblade's tile, would
make that first throw, "throw" it from the tile it is currently on
to that same tile. Which apparently causes it to go straight down.
Now, if the clickdrag starts from the blade's tile, the first throw is
updated to use the second throw's coordinates, and the second throw
already cancels itself if it targets its starting point.