Add the spiderbot fabricator
The spiderbot fabricator is a device built by R&D that can be activated by ghosts to spawn there as a spiderbot.
It is similar to the MoMMI fabricator, except it  must be fed metal sheets to work and can actually be built by players.
It can also be used by mice by clicking on it to become a mouse-spiderbot (which is apparently a thing).